+++ cmd/profiles.go
@@ -15,8 +15,8 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/elastic/elastic-package/internal/cobraext"
-	"github.com/elastic/elastic-package/internal/common"
 	"github.com/elastic/elastic-package/internal/configuration/locations"
+	"github.com/elastic/elastic-package/internal/environment"
 	"github.com/elastic/elastic-package/internal/profile"
 )
 
@@ -27,7 +27,7 @@ const jsonFormat = "json"
 const tableFormat = "table"
 
 // profileNameEnvVar is the name of the environment variable to set the default profile
-const profileNameEnvVar = common.ElasticPackageEnvPrefix + "PROFILE"
+const profileNameEnvVar = environment.WithElasticPackagePrefix("PROFILE")
 
 func setupProfilesCommand() *cobraext.Command {
 	profilesLongDescription := `Use this command to add, remove, and manage multiple config profiles.
